+++ apps/v4/app/(app)/(root)/components/button-group-demo.tsx
@@ -6,7 +6,7 @@ import {
   ArrowLeftIcon,
   CalendarPlusIcon,
   ClockIcon,
-  ListFilterPlusIcon,
+  ListFilterIcon,
   MailCheckIcon,
   MoreHorizontalIcon,
   TagIcon,
@@ -79,7 +79,7 @@ export function ButtonGroupDemo() {
                 Add to Calendar
               </DropdownMenuItem>
               <DropdownMenuItem>
-                <ListFilterPlusIcon />
+                <ListFilterIcon />
                 Add to List
               </DropdownMenuItem>
               <DropdownMenuSub>