+++ apps/v4/components/github-link.tsx
@@ -11,7 +11,7 @@ export function GitHubLink() {
     <Button asChild size="sm" variant="ghost" className="h-8 shadow-none">
       <Link href={siteConfig.links.github} target="_blank" rel="noreferrer">
         <Icons.gitHub />
-        <React.Suspense fallback={<Skeleton className="h-4 w-8" />}>
+        <React.Suspense fallback={<Skeleton className="h-4 w-[42px]" />}>
           <StarsCount />
         </React.Suspense>
       </Link>
@@ -21,15 +21,18 @@ export function GitHubLink() {
 
 export async function StarsCount() {
   const data = await fetch("https://api.github.com/repos/shadcn-ui/ui", {
-    next: { revalidate: 86400 }, // Cache for 1 day (86400 seconds)
+    next: { revalidate: 86400 },
   })
   const json = await data.json()
 
+  const formattedCount =
+    json.stargazers_count >= 1000
+      ? `${Math.round(json.stargazers_count / 1000)}k`
+      : json.stargazers_count.toLocaleString()
+
   return (
-    <span className="text-muted-foreground w-8 text-xs tabular-nums">
-      {json.stargazers_count >= 1000
-        ? `${(json.stargazers_count / 1000).toFixed(1)}k`
-        : json.stargazers_count.toLocaleString()}
+    <span className="text-muted-foreground w-fit text-xs tabular-nums">
+      {formattedCount}
     </span>
   )
 }